Sarancha_Lesson_6_work | Рефакторинг
